src/requirements.txt
astroid==2.1.0
attr==0.3.1
attrs==18.2.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pylev==1.3.0
pylint==2.2.2
python-dotenv==0.10.1
six==1.12.0
typed-ast==1.1.1
urllib3==1.24.1
wrapt==1.10.11
